Refactor workspace admin tests to fix the issue with the namespace
